update float to double and digit

+++ calc-devide.c
@@ -5,6 +5,6 @@ int main(void) {
   long x = get_long("X: ");
   long y = get_long("Y: ");
 
-  float res = (float) x / (float) y;
-  printf("Answer: %f\n", res);
+  double res = (double) x / (double) y;
+  printf("Answer: %.20f\n", res);
 }
